Move config functions to ovs_config.go, make getNetworkStateFromResources smarter for future recovery/restart.
